requirements.txt
google-cloud-aiplatform
python-dotenv
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
google-auth-oauthlib